refactor(agent-ui): split VMTable into modules and fix excluded-VMs toggle
- Break the monolithic VMTable into a hook, toolbar/grid/modals subcomponents,and shared variant/filter utilities.
- Add CompactVMTable for modal pickers,unit tests for filter chips and variants
- Fix the 'Show excluded VMs' toggle so it sends an explicit byExpression when enabled.
