+++ apps/agent-ui/src/pages/Report/components/index.ts
@@ -1,4 +1,5 @@
 export { ClustersOverview } from "./ClustersOverview";
+export { CompactVMTable } from "./CompactVMTable";
 export { CpuAndMemoryOverview } from "./CpuAndMemoryOverview";
 export { Dashboard } from "./Dashboard";
 export { DeepInspectionModal } from "./DeepInspectionModal";
