+++ apps/agent-ui/src/pages/Report/components/vmTableShared.ts
@@ -1,0 +1,186 @@
+import { css } from "@emotion/css";
+import type { VirtualMachine } from "@openshift-migration-advisor/agent-sdk";
+import { deepInspectionSort } from "./vmSort";
+
+export type ColumnKey =
+  | "name"
+  | "labels"
+  | "groups"
+  | "vCenterState"
+  | "id"
+  | "cpuUsage"
+  | "diskUsage"
+  | "ramUsage"
+  | "datacenter"
+  | "cluster"
+  | "diskSize"
+  | "memory"
+  | "issues"
+  | "migratable"
+  | "deepInspection";
+
+export const BACKEND_SORTABLE_COLUMNS = [
+  "name",
+  "vCenterState",
+  "cluster",
+  "diskSize",
+  "memory",
+  "issues",
+] as const;
+
+export const FRONTEND_SORTABLE_COLUMNS = ["deepInspection"] as const;
+
+export type BackendSortableColumn = (typeof BACKEND_SORTABLE_COLUMNS)[number];
+export type FrontendSortableColumn = (typeof FRONTEND_SORTABLE_COLUMNS)[number];
+export type SortableColumn = BackendSortableColumn | FrontendSortableColumn;
+
+export const Columns: Record<ColumnKey, string> = {
+  name: "Name",
+  labels: "Labels",
+  groups: "Groups",
+  vCenterState: "Status",
+  migratable: "Migration Readiness",
+  id: "ID",
+  cpuUsage: "CPU usage",
+  diskUsage: "Disk usage",
+  ramUsage: "RAM usage",
+  datacenter: "Data center",
+  cluster: "Cluster",
+  diskSize: "Disk size",
+  memory: "Memory size",
+  issues: "Issues",
+  deepInspection: "Deep inspection",
+};
+
+export const ALL_COLUMN_KEYS = Object.keys(Columns) as ColumnKey[];
+export const MANDATORY_COLUMNS: ColumnKey[] = ["name"];
+export const DEFAULT_VISIBLE_COLUMNS: ColumnKey[] = [...ALL_COLUMN_KEYS];
+
+export const COMPACT_VISIBLE_COLUMNS: ColumnKey[] = [
+  "name",
+  "labels",
+  "vCenterState",
+  "migratable",
+  "cluster",
+];
+
+export const VISIBLE_COLUMNS_KEY = "vmTable.visibleColumns";
+export const VISIBLE_COLUMNS_VERSION = 5;
+
+export const isSortableColumn = (key: ColumnKey): key is SortableColumn =>
+  (BACKEND_SORTABLE_COLUMNS as readonly ColumnKey[]).includes(key) ||
+  (FRONTEND_SORTABLE_COLUMNS as readonly ColumnKey[]).includes(key);
+
+export const isBackendSortableColumn = (
+  key: ColumnKey | null,
+): key is BackendSortableColumn =>
+  key !== null &&
+  (BACKEND_SORTABLE_COLUMNS as readonly ColumnKey[]).includes(key);
+
+export const statusLabels: Record<string, string> = {
+  poweredOn: "Powered on",
+  poweredOff: "Powered off",
+  suspended: "Suspended",
+};
+
+export const diskSizeRanges = [
+  { label: "0-10 TB", min: 0, max: 10 * 1024 * 1024 },
+  { label: "11-20 TB", min: 10 * 1024 * 1024 + 1, max: 20 * 1024 * 1024 },
+  { label: "21-50 TB", min: 20 * 1024 * 1024 + 1, max: 50 * 1024 * 1024 },
+  { label: "50+ TB", min: 50 * 1024 * 1024 + 1, max: undefined },
+];
+
+export const memorySizeRanges = [
+  { label: "0-4 GB", min: 0, max: 4 * 1024 },
+  { label: "5-16 GB", min: 4 * 1024 + 1, max: 16 * 1024 },
+  { label: "17-32 GB", min: 16 * 1024 + 1, max: 32 * 1024 },
+  { label: "33-64 GB", min: 32 * 1024 + 1, max: 64 * 1024 },
+  { label: "65-128 GB", min: 64 * 1024 + 1, max: 128 * 1024 },
+  { label: "129-256 GB", min: 128 * 1024 + 1, max: 256 * 1024 },
+  { label: "256+ GB", min: 256 * 1024 + 1, max: undefined },
+];
+
+const MB_IN_GB = 1024;
+const MB_IN_TB = 1024 * 1024;
+
+type FrontendSortFunction = (vm: VirtualMachine) => number;
+
+export const FRONTEND_SORT_METHODS: Record<
+  FrontendSortableColumn,
+  FrontendSortFunction
+> = {
+  deepInspection: deepInspectionSort,
+};
+
+export const formatDiskSize = (sizeInMB: number): string => {
+  if (sizeInMB >= MB_IN_TB) {
+    const sizeInTB = sizeInMB / MB_IN_TB;
+    return `${sizeInTB.toFixed(sizeInTB % 1 === 0 ? 0 : 2)} TB`;
+  }
+  const sizeInGB = sizeInMB / MB_IN_GB;
+  return `${sizeInGB.toFixed(sizeInGB % 1 === 0 ? 0 : 2)} GB`;
+};
+
+export const formatMemorySize = (sizeInMB: number): string => {
+  const sizeInGB = sizeInMB / MB_IN_GB;
+  return `${sizeInGB.toFixed(sizeInGB % 1 === 0 ? 0 : 2)} GB`;
+};
+
+export interface AppliedFilter {
+  category: string;
+  label: string;
+  key: string;
+}
+
+export const filterStyles = {
+  dropdownContent: css`
+    padding: 24px;
+    width: 1400px;
+    max-width: 95vw;
+    overflow: visible;
+  `,
+  filterGrid: css`
+    display: grid;
+    grid-template-columns: repeat(8, 1fr);
+    gap: 24px;
+  `,
+  concernsColumn: css`
+    max-height: 400px;
+    overflow-y: auto;
+  `,
+  concernSelect: css`
+    width: 100%;
+    margin-top: 8px;
+  `,
+  columnTitle: css`
+    font-size: 13px;
+    font-weight: 700;
+    margin-bottom: 16px;
+    color: var(--pf-t--global--text--color--regular);
+  `,
+  checkboxList: css`
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+  `,
+  footer: css`
+    display: flex;
+    justify-content: flex-start;
+    gap: 16px;
+    margin-top: 32px;
+    padding-top: 20px;
+  `,
+};
+
+export const vmTableStyles = {
+  vmTable: css`
+    th button {
+      display: flex;
+      align-items: center;
+      width: 100%;
+      text-align: left;
+      justify-content: space-between;
+      gap: 0.5rem;
+    }
+  `,
+};
